+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+Development v1
+
+### Changed
+
+- AVPlayer: current bitrate is now reported by `observedBitrate`
+
 ## 2.1.0
 
 ### Changed